refactor: http version entity validation for v072
